+++ app/builders/fallout_character/base_builder.rb
@@ -4,8 +4,8 @@ module FalloutCharacter
   class BaseBuilder
     def call(result:)
       result.merge({
-        attributes: { 'str' => 5, 'per' => 5, 'end' => 5, 'cha' => 5, 'int' => 5, 'agi' => 5, 'lck' => 5 },
-        max_attributes: {},
+        abilities: { 'str' => 5, 'per' => 5, 'end' => 5, 'cha' => 5, 'int' => 5, 'agi' => 5, 'lck' => 5 },
+        max_abilities: {},
         tag_skills: [],
         tag_skill_boosts: 3,
         guide_step: result[:skip_guide] ? nil : 1