+++ app/builders/fallout_character/origins/mutant_builder.rb
@@ -4,8 +4,8 @@ module FalloutCharacter
   module Origins
     class MutantBuilder
       def call(result:)
-        result[:attributes] = { 'str' => 7, 'per' => 5, 'end' => 7, 'cha' => 5, 'int' => 5, 'agi' => 5, 'lck' => 5 }
-        result[:max_attributes] = { 'str' => 12, 'end' => 12, 'cha' => 6, 'int' => 6 }
+        result[:abilities] = { 'str' => 7, 'per' => 5, 'end' => 7, 'cha' => 5, 'int' => 5, 'agi' => 5, 'lck' => 5 }
+        result[:max_abilities] = { 'str' => 12, 'end' => 12, 'cha' => 6, 'int' => 6 }
 
         result
       end