+++ app/serializers/fallout/character_serializer.rb
@@ -4,12 +4,13 @@ module Fallout
   class CharacterSerializer < ApplicationSerializer
     include Deps[cache: 'cache.avatars']
 
-    attributes :provider, :id, :name, :created_at, :avatar, :origin, :attributes, :carry_weight, :initiative, :health, :skills,
-               :tag_skills, :attibute_boosts, :tag_skill_boosts, :skill_boosts, :level
+    attributes :provider, :id, :name, :created_at, :avatar, :origin, :abilities, :carry_weight, :initiative, :health, :skills,
+               :tag_skills, :ability_boosts, :tag_skill_boosts, :skill_boosts, :level, :guide_step, :max_abilities
 
     delegate :carry_weight, :initiative, :health, to: :decorator
     delegate :data, to: :object
-    delegate :origin, :attributes, :skills, :tag_skills, :attibute_boosts, :tag_skill_boosts, :skill_boosts, :level, to: :data
+    delegate :origin, :abilities, :skills, :tag_skills, :ability_boosts, :tag_skill_boosts, :skill_boosts, :level, :guide_step,
+             :max_abilities, to: :data
 
     def provider
       'fallout'